Create data set for experiment
Used cnfgen to create the formulas that will be tested by our
implementation of the algorithm. These are stored in a compressed
zip file labeled "formulas.zip". I also included the script I used
to generate these files, and updated the .gitignore to include *.cnf
to avoid accidently adding the uncompressed files to repository.